Move `<markdown>` component to src/sidebar/components
